+++ docs/conf.py
@@ -258,7 +258,7 @@ latex_elements = {
 # Grouping the document tree into LaTeX files. List of tuples
 # (source start file, target name, title, author, documentclass [howto/manual]).
 latex_documents = [
-    ("index", "user_guide.tex", u"PICor Documentation", u"Joern Dietze", "manual"),
+    ("index", "user_guide.tex", u"PICor Documentation", u"Joern Dietze", "manual")
 ]
 
 # The name of an image file (relative to this directory) to place at the top of